backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
tavily-python==0.7.21
python-dotenv==1.1.1
requests==2.32.5
beautifulsoup4==4.13.5
openai==1.108.1
